requirements.txt
faiss-cpu==1.8.0.post1
numpy==1.26.4
requests==2.32.3
sentence-transformers==3.0.1
torch==2.3.1
transformers==4.42.4
